server/requirements.txt
Flask==2.3.1
Flask-Cors==3.0.9
mariadb==1.1.5
requests==2.19.1
pycryptodome==3.17
pandas==2.2.2
openpyxl==3.1.5
PyJWT==2.8.0
cryptography==41.0.3
